Svace fix for cloud Client functionality
Reviewed-on: https://gerrit.iotivity.org/gerrit/12223
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
(cherry picked from commit 0ca71f3fcca638bdd199bfe7c856fd4da08e7822)
Reviewed-on: https://gerrit.iotivity.org/gerrit/12321
Tested-by: Randeep Singh <randeep.s@samsung.com>

+++ resource/csdk/security/provisioning/src/cloud/aclinvite.c
@@ -35,7 +35,7 @@ static OCStackResult parseInvitePayload(const OCRepPayload *payload, const char 
         return OC_STACK_MALFORMED_RESPONSE;
     }
 
-    int count = calcDimTotal(dimensions);
+    size_t count = calcDimTotal(dimensions);
 
     stringArray_t *gidlist = &out->gidlist;
     stringArray_t *midlist = &out->midlist;
